requirements.txt
# Core CLI & Config
python-dotenv==1.0.0
typer==0.9.0
click==8.1.7
rich==13.7.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# LLM (Local Gemma via Ollama)
ollama==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Data & Analysis
pandas>=2.0.0
numpy>=1.26.0
ta==0.11.0

# Document Processing
pypdf==4.0.0
python-docx==0.8.11
markdown==3.5.1
python-frontmatter==1.1.1

# Utilities
python-dateutil==2.8.2
requests==2.31.0
schedule==1.2.0

# Optional: MT5 (Windows only, uncomment if available)
# MetaTrader5==5.0.45
